Trap SIGHUP and SIGPIPE.
Trap SIGHUP and SIGPIPE to prevent the script from dieing mid-migration due to a
remote connection dropping out.  If this were to happen then it would leave the system in a highly unstable state and it's better to allow the script to finish behind the scenes than to kill it at this stage.